+++ app/types/index.ts
@@ -35,6 +35,7 @@ export interface TemplateInterface {
   name: string;
   description?: string | null;
   modified?: string | null;
+  modifiedById?: number | null;
   id?: number | null;
   isDirty?: boolean;
   owner?: { // Make `owner` optional and nullable
@@ -42,8 +43,30 @@ export interface TemplateInterface {
     name?: string;
     searchName?: string;
   } | null;
+  visibility?: string;
 }
 
+export interface MyVersionedTemplatesInterface {
+  id?: number | null;
+  name: string;
+  description?: string | null;
+  modified?: string | null;
+  modifiedById?: number | null;
+  versionType?: string | null;
+  visibility: string;
+  errors?: string[] | null;
+  template?: {
+    __typename?: string; // Match GraphQL's optional __typename
+    id?: number | null;
+    owner?: {
+      __typename?: string; // Match GraphQL's optional __typename
+      id?: number | null;
+      displayName?: string; // Make displayName optional
+      name?: string; // Make name optional
+      searchName?: string; // Make searchName optional
+    } | null;
+  } | null; // Allow `template` to be null or undefined
+}
 // Define valid href types for Next.js Link
 type Url = string | URL;
 type LinkHref = Url | {
@@ -53,13 +76,18 @@ type LinkHref = Url | {
 };
 
 export interface TemplateItemProps {
+  id?: number | null;
+  template?: {
+    id?: number | null;
+  },
   title: string;
   content?: JSX.Element | null;
+  description?: string;
   link?: LinkHref;
   defaultExpanded: boolean;
   funder?: string;
   lastUpdated?: string | null;
-  publishStatus?: string;
+  publishStatus?: string | null;
 }
 
 export interface TemplateVersionInterface {