+++ lib/graphql/apolloClient.ts
@@ -1,0 +1,10 @@
+// Create a new file: apolloClient.ts
+import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
+
+let apolloClient: ApolloClient<NormalizedCacheObject> | null = null;
+
+export const setApolloClient = (client: ApolloClient<NormalizedCacheObject>) => {
+  apolloClient = client;
+};
+
+export const getApolloClient = () => apolloClient;