+++ CHANGELOG.md
@@ -1,8 +1,38 @@
+## v0.0.1
+
+### Updated
+- Updated `/template/create/page.tsx`[#186]
+- Removed select-template page, since we are using `SelectExistingTemplate` in place of it [#186]
+- Updated existing Templates graphql query with more fields [#186]
+- Updated shared `FormInput` component to pass ‘ariaDescribedBy’ [#186]
+- Updated `TemplateSelectListItem` [#186]
+- Updated `/graphql/apollo-wrapper.tsx` since `@apollo/experimental-nextjs-app-support/ssr` was deprecated [#186]
+- Updated `/graphql/graphqlHelper.ts` with a call to `client.resetStore’` This should refetch user’s data when refreshAuthTokens() is called [#186]
+- Updated `refreshAuthTokens` function with ability to pass in cookies [#186]
+- Updated `middleware` to call refreshAuthTokens when there is no auth token cookie [#186]
+- Updated `QuestionTypeSelectPage`- hooked up to real data and updated code [#220]
+- Renamed previous `QuestionEdit` component to `QuestionEditCard` for more specificity [#220]
+
+### Added
+- Added new `QuestionEdit` and `QuestionTypeCard` components [#220]
+
+
+### Fixed
+
+
+====================================================================================================
 
 ### Updated
 =======
-- Updated `QuestionTypeSelectPage`- hooked up to real data and updated code [#220]
-- Renamed previous `QuestionEdit` component to `QuestionEditCard` for more specificity [#220]
+- Updated `/template/create/page.tsx`[#186]
+- Removed select-template page, since we are using `SelectExistingTemplate` in place of it [#186]
+- Updated existing Templates graphql query with more fields [#186]
+- Updated shared `FormInput` component to pass ‘ariaDescribedBy’ [#186]
+- Updated `TemplateSelectListItem` [#186]
+- Updated `/graphql/apollo-wrapper.tsx` since `@apollo/experimental-nextjs-app-support/ssr` was deprecated [#186]
+- Updated `/graphql/graphqlHelper.ts` with a call to `client.resetStore’` This should refetch user’s data when refreshAuthTokens() is called [#186]
+- Updated `refreshAuthTokens` function with ability to pass in cookies [#186]
+- Updated `middleware` to call refreshAuthTokens when there is no auth token cookie [#186]
 - Updated some pages with toast messages, and updated toast files [#231]
 - Added Portuguese translations for newly added translation keys [#231]
 - Updated version of `next` to `14.2.22` [#231]
@@ -16,7 +46,9 @@
 - Refactored font family usage for consistency.
 
 ### Added
-- Added new `QuestionEdit` and `QuestionTypeCard` components [#220]
+- Added `MyVersionedTemplates` and `PublishedTemplates` graphql queries and `addTemplateMutation` graphql mutation [#186]
+- Created the `SelectExistingTemplate` component which displays the publishedTemplates and allows user to filter view [#186]
+- Created `TemplateList` component used by `SelectExistingTemplate`[#186]
 - Added Toast Message capabilities using **React Aria Component's use Toast**. [#211]
 - Project overview page [#175]
   - Moved pages under [locale] folder